+++ os_in_browser/src/app/Finder/index.tsx
@@ -1,9 +1,14 @@
 import SideBar from "./sidebar";
 import FinderMain from "./finderMain";
+import { createSignal, createEffect } from "solid-js";
 
 import { FileDirectoryProvider } from "../../context/FileDirectoryContext";
-
+import { mainFileSystem } from "../../sharedHooks/useFileSystem";
 const Finder = () => {
+  const [cd, setCd] = createSignal("");
+
+  createEffect(() => {});
+
   return (
     <FileDirectoryProvider>
       <div class="box-body">
